td-shim: split tcg.rs into two files
Split tcg.rs into two files to solve an `asm!()` related compilation
failure.
